QVAC-19247 feat[api]: expose @qvac/rag/errors subpath for consumers (#2303)
Add an exports field to @qvac/rag and ship src/errors.d.ts so consumers
(notably @qvac/sdk) can import ERR_CODES and QvacErrorRAG from
@qvac/rag/errors without dragging in the full main entry, which eagerly
loads HyperDBAdapter and breaks bare-pack consumer bundles.

index.d.ts now re-exports these from ./src/errors, making the new file
the single .d.ts source of truth for the error codes/class.
